requirements.txt
yt-dlp
openai-whisper
ffmpeg-python
pydub
